Initial commit and findings
Assets added and build target configured to a horizontal iPad. So far I've observed that the iPad simulator takes a bit to switch from portrait to landscape. I also noticed in the assets folder I want integration with Apple's sticker generator. This might already be in Xcode 26. Lastly I want a way to add a border to a label and make its background transparent without reducing the text's transparency a it is making the label difficult to read due to the low contrast. This may be fixed by liquid glass.